Refactor and customize cli for claimable nft queries
Ref oursky/likecoin-chain#202
