--- x/likenft/client/cli/query_claimable_nfts.go
@@ -1,52 +1,0 @@
-package cli
-
-import (
-	"strconv"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/likecoin/likechain/x/likenft/types"
-	"github.com/spf13/cobra"
-)
-
-var _ = strconv.Itoa(0)
-
-func CmdClaimableNFTs() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "claimable-nfts [class-id]",
-		Short: "Query ClaimableNFTs",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			reqClassId := args[0]
-
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			params := &types.QueryClaimableNFTsRequest{
-
-				ClassId: reqClassId,
-			}
-
-			pageReq, err := client.ReadPageRequest(cmd.Flags())
-			if err != nil {
-				return err
-			}
-			params.Pagination = pageReq
-
-			res, err := queryClient.ClaimableNFTs(cmd.Context(), params)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
